+++ trixnity-messenger-compose-view/src/webMain/kotlin/de/connect2x/messenger/compose/view/Notifications.kt
@@ -150,14 +150,12 @@ private suspend fun displayNotification(
                 user?.name to image
             } ?: (null to null)
 
-            val title = if (isDirect) username.orEmpty() else roomName
-            val text = if (isDirect) message else "$username: $message"
-
             log.debug { "notification will appear" }
 
             return Notification(
-                title = title,
-                description = text,
+                title = if (isDirect) username.orEmpty() else roomName,
+                group = "${matrixClient.userId}-$roomId",
+                description = if (isDirect) message else "$username: $message",
                 icon = imageInBytes?.let { NotificationIcon(it, avatarSize(), avatarSize()) })
         }
     } ?: log.warn { "cannot find roomId for event ${event.idOrNull}" }